Move getQuestion to collab room
Pass question into question display instead.
Use Question schema instead of creating a new interface
